refactoring pagination
