+++ src/js/search.js
@@ -6,7 +6,7 @@ import { createPagination } from './pagination';
 const searchForms = document.querySelectorAll('.search-form');
 const contentResults = document.querySelector('#content-results');
 const contentTitle = document.querySelector('#content-title');
-const pagination = document.querySelector('#pagination');
+
 
 const mainSearch = searchForms[0];
 const mobSearch = searchForms[1];
@@ -21,7 +21,6 @@ mainSearch.addEventListener('input', () => {
 mobSearch.addEventListener('input', () => {
   mainSearch.elements.search.value = mobSearch.elements.search.value;
 });
-let searchResults;
 
 function getCocktails(event) {
   event.preventDefault();
@@ -33,18 +32,16 @@ function getCocktails(event) {
 
   fetchCocktailsByName(cocktailName)
     .then(data => {
-      searchResults = data;
       contentTitle.innerHTML = 'Cocktails';
       contentResults.innerHTML = "";
       contentResults.append ( ...getDrinksMarkup(
         data.slice(0, numberOfCocktails)
       ));
-      pagination.innerHTML = createPagination(
-        data.length,
+      createPagination(
+        data,
         numberOfCocktails,
         1
       );
-      addPaginationListeners();
     })
     .catch((error) => {
       contentTitle.innerHTML = error
@@ -52,23 +49,4 @@ function getCocktails(event) {
     });
 }
 
-function changePage(e) {
-  const page = Number(e.target.dataset.page);
-  const numberOfCocktails = calculateCocktails();
-  const data = searchResults;
-  const end = page * numberOfCocktails;
-  const start = end - numberOfCocktails;
-  contentResults.innerHTML = '';
-  contentResults.append (...getDrinksMarkup(data.slice(start, end)));
-  pagination.innerHTML = createPagination(data.length, numberOfCocktails, page);
-  addPaginationListeners();
-}
 
-function addPaginationListeners() {
-  const pageElements = document.querySelectorAll('.pagination__number');
-  for (const el of pageElements) {
-    if (el.className.includes('inactive')) continue;
-    el.addEventListener('click', changePage);
-  }
-
-}
